# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Application API",
        "description": "API Documentation for Your Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/scholarships": {
            "get": {
                "tags": [
                    "Scholarships"
                ],
                "summary": "Get all scholarships",
                "description": "Display a listing of the scholarships.",
                "operationId": "0ee5b3990cc5093347f3a22a8b655ef7",
                "responses": {
                    "200": {
                        "description": "List of scholarships",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "eligibility": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "criteria": {
                                                            "type": "string"
                                                        },
                                                        "minimum_gpa": {
                                                            "type": "number"
                                                        },
                                                        "country": {
                                                            "type": "string"
                                                        },
                                                        "experience": {
                                                            "type": "integer"
                                                        },
                                                        "english_proficiency": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Scholarships"
                ],
                "summary": "Create a new scholarship",
                "description": "Create a new scholarship with eligibility criteria",
                "operationId": "5264958941033b303a3899a554b81c21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "eligibility"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Scholarship Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Scholarship Description"
                                    },
                                    "eligibility": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "criteria": {
                                                    "type": "string",
                                                    "example": "Eligibility Criteria"
                                                },
                                                "minimum_gpa": {
                                                    "type": "number",
                                                    "example": 3.5
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "Country Name"
                                                },
                                                "experience": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "english_proficiency": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Scholarship created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "eligibility": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "criteria": {
                                                        "type": "string"
                                                    },
                                                    "minimum_gpa": {
                                                        "type": "number"
                                                    },
                                                    "country": {
                                                        "type": "string"
                                                    },
                                                    "experience": {
                                                        "type": "integer"
                                                    },
                                                    "english_proficiency": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}